ENHANCEMENT: Rename PayPal Gateway add-on module ENHANCEMENT: Add OAuth2 Client Secret and Key settings
